Add support for local PoW
To use it, call the `localPoW` method of the IotaAPI.Builder and pass
your desired PoW implementation.

The included implementation uses @Come-from-Beyond's PearlDiver (which
according to a statement of him on Slack may be used "for whatever you
wish" despite missing a license in the repo) to performs PoW by CPU.
Other implementations are possible (even outside this lib) by
implementing the IotaLocalPoW interface.

The PearlDiver class has been taken from
iotaledger/PearlDiver@212c332d6002095e9dc8f74789deb3747b8b4870, the only
modification was to make it Java 7 compatible by replacing the lambda
with a Runnable anonymous class and fixing the nested `this` references
to point to the outer class.